requirements.txt
playwright
pytest
pytest-playwright
pytest-html
allure-pytest
openpyxl
pandas
